feat: implement infinite scroll for news stories and add intersection observer
