+++ apps/web/package.json
@@ -22,6 +22,7 @@
 		"react-dom": "^19.0.0",
 		"react-hook-form": "^7.54.2",
 		"react-icons": "^5.4.0",
+		"react-intersection-observer": "^9.16.0",
 		"sass": "^1.85.0",
 		"sdk": "link:@/ory/sdk",
 		"serwist": "^9.0.11",
